FIX: Not scaling the data is a bad default.
If the user of read_pdata tries to compare two spectrum and does not
scale the data the result will normally be incorrect. If read and
read_pdata are used on the same dataset, to get the same result the
data needs to be scaled. I think that it should not even be an option
to scale the data.